--- a/nim.cfg
+++ b/nim.cfg
@@ -11,6 +11,10 @@
   --passL:"-Wl,--stack,8388608"
   # https://github.com/nim-lang/Nim/issues/4057
   --tlsEmulation:off
+  @if i386:
+    # set the IMAGE_FILE_LARGE_ADDRESS_AWARE flag so we can use PAE, if enabled, and access more than 2 GiB of RAM
+    --passL:"-Wl,--large-address-aware"
+  @end
 @end
 
 --threads:on