--- a/tests/all_tests.nim
+++ b/tests/all_tests.nim
@@ -5,7 +5,11 @@
 #  * MIT license ([LICENSE-MIT](LICENSE-MIT) or http://opensource.org/licenses/MIT)
 # at your option. This file may not be copied, modified, or distributed except according to those terms.
 
-import macros, strutils, os, unittest, osproc
+import macros, strutils, os, unittest2, osproc
+import threadpool
+
+# AppVeyor may go out of memory with the default of 4
+setMinPoolSize(2)
 
 proc executeMyself(numModules: int): int =
   let appName = getAppFilename()