--- a/tests/nim.cfg
+++ b/tests/nim.cfg
@@ -1,3 +1,5 @@
 -d:chronicles_line_numbers
 -d:"chronicles_sinks=textblocks"
+# comment this out, to run the tests in a serial manner:
+#-d:nimtestParallel