--- a/tests/test_generalstate_json.nim
+++ b/tests/test_generalstate_json.nim
@@ -6,7 +6,7 @@
 # at your option. This file may not be copied, modified, or distributed except according to those terms.
 
 import
-  unittest, strformat, strutils, tables, json, ospaths, times, os,
+  unittest2, strformat, strutils, tables, json, ospaths, times, os,
   stew/byteutils, stew/ranges/typedranges, nimcrypto, options,
   eth/[rlp, common], eth/trie/[db, trie_defs], chronicles,
   ./test_helpers, ../nimbus/p2p/executor, test_config,