--- a/tests/test_genesis.nim
+++ b/tests/test_genesis.nim
@@ -1,4 +1,4 @@
-import unittest, ../nimbus/[genesis, config], eth/common, nimcrypto/hash
+import unittest2, ../nimbus/[genesis, config], eth/common, nimcrypto/hash
 
 proc genesisMain*() =
   suite "Genesis":