--- a/tests/test_memory.nim
+++ b/tests/test_memory.nim
@@ -6,7 +6,7 @@
 # at your option. This file may not be copied, modified, or distributed except according to those terms.
 
 import
-  unittest, sequtils,
+  unittest2, sequtils,
   eth/common/eth_types,
   ../nimbus/[constants, errors, vm/memory]