--- a/tests/test_op_arith.nim
+++ b/tests/test_op_arith.nim
@@ -1,8 +1,9 @@
-import macro_assembler, unittest
+import macro_assembler, unittest2
 
 proc opArithMain*() =
   suite "Arithmetic Opcodes":
-    let (blockNumber, chainDB) = initDatabase()
+    setup:
+      let (blockNumber, chainDB) = initDatabase()
 
     assembler:
       title: "ADD_1"