--- a/tests/test_op_bit.nim
+++ b/tests/test_op_bit.nim
@@ -1,8 +1,9 @@
-import macro_assembler, unittest
+import macro_assembler, unittest2
 
 proc opBitMain*() =
   suite "Bitwise Opcodes":
-    let (blockNumber, chainDB) = initDatabase()
+    setup:
+      let (blockNumber, chainDB) = initDatabase()
 
     assembler: # AND OP
       title: "AND_1"