@jangko 2019-05-31
* new entries in `PersistBlockTests.md` and `TracerTests.md` are positively incorrect. old entries already correct.
* `PrecompileTests.md` also contains suspiciously similar entries like the two above, I assume it also incorrect.

I believe the bug is in `tests/test_helpers.nim`, perhaps a subtle non thread safe obscured somewhere.