@stefantalpalaru 2019-05-31
> I believe the bug is in tests/test_helpers.nim

It was: I made the `status` table a global var, then forgot to clear it between test suites.