@jlokier 2021-05-26
Times have changed - now the testsuite is huge, and takes longer to run than compiling.  The original patch won't apply any more, but the idea is welcome.  Parallel tests would be great!

(Mentioning it here because the PR is still open. Might as well use it or close it.)

If there is an interest in bringing this change or another like it up to date, the comments below which discuss other things we'd like may be relevant (from https://github.com/status-im/nimbus-eth1/pull/683#issuecomment-849043802):

> .. capture log output for each test individually and save it to a file (or xml/json report) so it can be viewed per test - specially per failed test. This requires fixing the test runner, rather than changing the production code.  Finally, one could consider not printing anything for tests that succeeded - that typically requires a more sophisticated test runner that can show a progress bar or something like that - it would simply print fails.

> If there ever is a more sophisticated test runner that captures output to hide it on success, it would be good to run tests in parallel too, and use the capture functionality to serialise their outputs, both to the log and to the terminal. I'm pretty sure there's a tool for this, but I don't recall its name.