@arnetheduck 2021-05-27
the (easy) way to run tests in parallel is to run them as separate processes - it doesn't require anything of the testing framework and makes sure things are nicely isolated - ie this is specially important because we don't use threading a lot and some of the libraries use globals and other thread-incompatible constructs that don't get tested a lot. 

That said, transitioning to unittest2 should be done regardless - its biggest advantage is that it puts each test in a separate `proc` which gives it a cleaner environment and stack, and doesn't pollute the global table in nim which is limited. It's also in ut2 that we can add features such as capturing output easily.

Finally, @jangko was looking into mocha - writing a mocha collector that outputs the json (without using json-serialization) would also be done in unittest2.