@jlokier 2021-05-27
> the (easy) way to run tests in parallel is to run them as separate processes

I agree.  There are some downsides to performance but it's much better isolation, and things that _might_ use globals, threadvars, timers, etc. need that isolation for the tests to mean anything.  (And some tests would want to launch their own threads - e.g. threadpool tests).

Processes also solve the stdout/stderr-are-not-per-thread problem.

On unix `fork` is reasonably cheap, but not ultra-cheap.  `fork+exec` is how you spawn a program normally, but that's less cheap, partly the `exec`, and partly program initialisations.  Keep in mind we have some ~12500 or so tests currently, and we'd like to run another ~28000, so every 1ms adds significantly to the testing time.

Windows doesn't have `fork`.  Cygwin simulates it in a complicated way, but not only are we not using Cygwin, Cygwin `fork` is slow anyway.  The only practical choice on Windows is equivalent to unix `fork+exec`.

Between those two, I think there's merit in using a "processpool" to allow related tests to run in the same process sequentially, for tests we're reasonably confident are fine with that.