@stefantalpalaru 2021-05-27
> transitioning to unittest2

Most of Nimbus-eth1 tests already use unittest2. Only these are still on unittest:

```text
tests/test_rpc.nim
tests/rpcclient/test_hexstrings.nim
tests/test_rpc_whisper.nim
tests/test_graphql.nim
```