@jlokier 2021-05-27
If it's using unittest2, and the headline feature is parallelism, why don't I see any?  It's 1 core all the way.