@arnetheduck 2021-05-27
> the headline feature is parallelism

it's not, any more - it has to be enabled, and is not working reliably - there are other features in there however that are significant, in particular `proc` isolation for each test (fewer globals)